Delete all code besides variables to restart
I am doing this because I have realized I need to
do a loop within a loop, as the issue with previous
code is that the num[i] value is not properly checking
against the array[i] value given the different sizes
of the array lengths and how the if boolean checks work.

Basically, I need to have a for loop cycle through each of the num
array values and properly check that against every single array value
before continuing to the next array value.

I think it will be easier to start from scratch (less einstellung).