Add issue for bug in comments
The bug likely has to do with array.splice(num[x] - 1, 1)
given with multi value entries it is bugging out...

I know! Its because when the value is spliced it removes
it from the check so the next array[i] check is against
the updated array. The solution should be in first
returning a new array, then after the loops just
updating array to match the new array so there are no
issues.